+++ src/main/java/seedu/address/logic/commands/ExportCommand.java
@@ -14,7 +14,7 @@ import javafx.collections.ObservableList;
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.util.ExportUtil;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.FilePath;
+import seedu.address.model.export.FilePath;
 import seedu.address.model.Model;
 import seedu.address.model.category.Category;
 import seedu.address.model.category.CategoryContainsAnyKeywordsPredicate;
