+++ src/main/java/seedu/address/model/export/FilePath.java
@@ -1,6 +1,6 @@
 //@@author LeowWB
 
-package seedu.address.model;
+package seedu.address.model.export;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
