+++ src/test/java/seedu/address/model/FilePathTest.java
@@ -5,6 +5,8 @@ import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.model.export.FilePath;
+
 public class FilePathTest {
 
     @Test
